# swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TodoListApp API",
    "description": "사용자 인증 기반의 개인화된 할일 관리 서비스 REST API. JWT Bearer Token을 통한 인증이 필요하며, 사용자별 데이터는 완전히 격리됩니다.",
    "version": "1.0.0",
    "contact": {
      "name": "개발팀"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1 서버"
    }
  ],
...
  "paths": {
...
    "/todos/{id}": {
      "get": {
        "tags": ["Todos"],
        "summary": "할일 상세 조회",
        "description": "지정한 ID의 할일 상세 정보를 조회합니다.",
        "operationId": "getTodoById",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TodoId"
          }
        ],
        "responses": {
          "200": {
            "description": "할일 상세 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                },
                "example": {
                  "id": 456,
                  "userId": 123,
                  "categoryId": 5,
                  "title": "클라이언트C 웹사이트 리디자인 - UI 프로토타입",
                  "description": "클라이언트C의 신규 웹사이트 리디자인 프로젝트.",
                  "dueDate": "2026-05-31",
                  "isCompleted": false,
                  "createdAt": "2026-05-13T14:30:45.123Z",
                  "updatedAt": "2026-05-13T14:30:45.123Z"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "존재하지 않는 할일",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "TODO_001",
                    "message": "할일을 찾을 수 없습니다"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "patch": {
        "tags": ["Todos"],
        "summary": "할일 수정",
        "description": "지정한 할일의 정보를 수정합니다. 변경할 필드만 포함하여 요청합니다. (부분 업데이트)",
        "operationId": "updateTodo",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TodoId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTodoRequest"
              },
              "examples": {
                "카테고리 변경": {
                  "value": {
                    "category_id": 6
                  }
                },
                "전체 수정": {
                  "value": {
                    "title": "수정된 할일 제목",
                    "description": "수정된 설명",
                    "category_id": 2,
                    "due_date": "2026-06-01"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "할일 수정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                },
                "example": {
                  "id": 150,
                  "userId": 123,
                  "categoryId": 6,
                  "title": "클라이언트D 웹개발 - 프론트엔드 구현",
                  "description": null,
                  "dueDate": "2026-05-25",
                  "isCompleted": false,
                  "createdAt": "2026-05-10T09:00:00Z",
                  "updatedAt": "2026-05-13T15:30:00Z"
                }
              }
            }
          },
          "400": {
            "description": "유효성 검사 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "VALID_001",
                    "message": "필수 입력 항목 누락 또는 형식 오류"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "존재하지 않는 할일 또는 카테고리",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "할일 없음": {
                    "value": {
                      "error": {
                        "code": "TODO_001",
                        "message": "할일을 찾을 수 없습니다"
                      }
                    }
                  },
                  "카테고리 없음": {
                    "value": {
                      "error": {
                        "code": "CAT_001",
                        "message": "카테고리를 찾을 수 없습니다"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": ["Todos"],
        "summary": "할일 삭제",
        "description": "지정한 할일을 삭제합니다.",
        "operationId": "deleteTodo",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TodoId"
          }
        ],
        "responses": {
          "200": {
            "description": "할일 삭제 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                },
                "example": {
                  "message": "할일이 삭제되었습니다"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "존재하지 않는 할일",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "TODO_001",
                    "message": "할일을 찾을 수 없습니다"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Access Token. 로그인 후 발급된 accessToken을 입력하세요. 유효 시간: 1시간"
      }
    },
    "parameters": {
      "TodoId": {
        "name": "id",
        "in": "path",
        "description": "할일 ID",
        "required": true,
        "schema": {
          "type": "integer",
          "example": 456
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "인증 오류 (Authorization 헤더 없음, Access Token 만료 또는 무효)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "헤더 없음": {
                "value": {
                  "error": {
                    "code": "AUTH_001",
                    "message": "Authorization 헤더가 없습니다"
                  }
                }
              },
              "토큰 만료": {
                "value": {
                  "error": {
                    "code": "AUTH_002",
                    "message": "Access Token이 만료되었습니다"
                  }
                }
              },
              "유효하지 않은 토큰": {
                "value": {
                  "error": {
                    "code": "AUTH_003",
                    "message": "유효하지 않은 토큰입니다"
                  }
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "타인의 리소스에 접근 시도",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "AUTH_004",
                "message": "접근 권한이 없습니다"
              }
            }
          }
        }
      },
      "InternalServerError": {
        "description": "내부 서버 오류",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "SERVER_001",
                "message": "내부 서버 오류가 발생했습니다"
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "Todo": {
        "type": "object",
        "description": "할일 엔티티",
        "properties": {
          "id": {
            "type": "integer",
            "description": "할일 고유 ID",
            "example": 456
          },
          "userId": {
            "type": "integer",
            "description": "할일 소유 사용자 ID",
            "example": 123
          },
          "categoryId": {
            "type": "integer",
            "description": "할일이 속한 카테고리 ID",
            "example": 5
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "description": "할일 제목 (필수, 최대 200자)",
            "example": "클라이언트C 웹사이트 리디자인 - UI 프로토타입"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true,
            "description": "할일 상세 설명 (선택, 최대 2000자)",
            "example": "클라이언트C의 신규 웹사이트 리디자인 프로젝트."
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "description": "종료 예정일 (YYYY-MM-DD 형식)",
            "example": "2026-05-31"
          },
          "isCompleted": {
            "type": "boolean",
            "description": "완료 여부 (기본값: false)",
            "example": false,
            "default": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "생성 일시 (UTC)",
            "example": "2026-05-13T14:30:45.123Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "수정 일시 (UTC)",
            "example": "2026-05-13T14:30:45.123Z"
          }
        },
        "required": ["id", "userId", "categoryId", "title", "dueDate", "isCompleted", "createdAt", "updatedAt"]
      },
...
      "ErrorDetail": {
        "type": "object",
        "description": "에러 상세 정보",
        "properties": {
          "code": {
            "type": "string",
            "description": "에러 코드",
            "enum": [
              "AUTH_001",
              "AUTH_002",
              "AUTH_003",
              "AUTH_004",
              "AUTH_005",
              "USER_001",
              "USER_002",
              "TODO_001",
              "TODO_002",
              "CAT_001",
              "CAT_002",
              "CAT_003",
              "VALID_001",
              "SERVER_001"
            ],
            "example": "AUTH_001"
          },
          "message": {
            "type": "string",
            "description": "에러 설명 메시지",
            "example": "Authorization 헤더가 없습니다"
          }
        },
        "required": ["code", "message"]
      },
      "ErrorResponse": {
        "type": "object",
        "description": "에러 응답 공통 형식",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": ["error"]
      },
      "MessageResponse": {
        "type": "object",
        "description": "단순 메시지 응답",
        "properties": {
          "message": {
            "type": "string",
            "description": "처리 결과 메시지",
            "example": "처리가 완료되었습니다"
          }
        },
        "required": ["message"]
      },
...
      "UpdateTodoRequest": {
        "type": "object",
        "description": "할일 수정 요청 (변경할 필드만 포함, 부분 업데이트)",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "description": "변경할 할일 제목 (최대 200자)",
            "example": "수정된 할일 제목"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true,
            "description": "변경할 할일 설명 (최대 2000자)",
            "example": "수정된 설명"
          },
          "category_id": {
            "type": "integer",
            "description": "변경할 카테고리 ID",
            "example": 2
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "description": "변경할 종료 예정일 (YYYY-MM-DD 형식)",
            "example": "2026-06-01"
          }
        }
      }
    }
  }
}